moved some files to data folder
