+++ iracing_season_organizer/app.py
@@ -5,7 +5,8 @@ from .data.free_content import get_free_content
 from .xlsx.cell import set_cell_styles
 from .xlsx.legend import print_buttons, print_owned_missing, print_classes
 import xlsxwriter
-from .dics import content, col_sizes
+from .data.dics import col_sizes
+from .data.colors import content
 from .xlsx.ir_license import get_license_colors
