+++ iracing_season_organizer/data/dics.py
@@ -1,0 +1,6 @@
+col_sizes = {
+    'road': {},
+    'oval': {},
+    'dirt_road': {},
+    'dirt_oval': {}
+}
